+++ wear/src/main/kotlin/com/auriqa/music/wear/ui/MainActivity.kt
@@ -5,7 +5,6 @@ import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.ui.Modifier
-import androidx.wear.compose.material.MaterialTheme
 import androidx.wear.compose.material.Scaffold
 import androidx.wear.compose.material.TimeText
 import com.auriqo.music.wear.media.AuriqoMediaSessionService
@@ -18,9 +17,7 @@ class MainActivity : ComponentActivity() {
         AuriqoMediaSessionService.start(this)
 
         setContent {
-            MaterialTheme(
-                colors = AuriqoWearColors.themeColors,
-            ) {
+            AuriqoWearColors.Theme {
                 Scaffold(
                     timeText = { TimeText() },
                 ) {